Move isAuthorized to its own module
